+++ j4ccl-ssh/src/main/java/com/cerner/ccl/j4ccl/impl/util/AuthHelper.java
@@ -7,7 +7,9 @@ import javax.security.auth.login.Configuration;
 import com.cerner.ccl.j4ccl.impl.jaas.BackendNodePasswordCredential;
 import com.cerner.ccl.j4ccl.impl.jaas.BackendNodePrincipal;
 import com.cerner.ccl.j4ccl.impl.jaas.JaasUtils;
+import com.cerner.ccl.j4ccl.impl.jaas.PrivateKeyPrincipal;
 import com.cerner.ftp.data.FtpProduct;
+import com.cerner.ftp.data.sftp.KeyCryptoBuilder;
 import com.cerner.ftp.data.sftp.UserPassBuilder;
 
 /**
@@ -34,7 +36,12 @@ public final class AuthHelper {
         final BackendNodePrincipal principal = JaasUtils.getPrincipal(BackendNodePrincipal.class);
         final BackendNodePasswordCredential credential = JaasUtils
                 .getPrivateCredential(BackendNodePasswordCredential.class);
-
+        if (JaasUtils.hasPrincipal(PrivateKeyPrincipal.class)) {
+            PrivateKeyPrincipal keyPrincipal = JaasUtils.getPrincipal(PrivateKeyPrincipal.class);
+            return KeyCryptoBuilder.getBuilder().setKeySalt(credential.getPassword())
+                    .setPrivateKey(URI.create(keyPrincipal.getName()))
+                    .setServerAddress(URI.create(principal.getHostname())).setUsername(principal.getUsername()).build();
+        }
         return UserPassBuilder.getBuilder().setServerAddress(URI.create(principal.getHostname()))
                 .setUsername(principal.getUsername()).setPassword(credential.getPassword()).build();
     }